# docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: Enclii Platform API
  description: |
    Open source DevOps platform API for deploying and managing containerized applications.

    ## Overview

    Enclii Switchyard is the control plane API that manages projects, services, deployments,
    domains, teams, and infrastructure. It provides a comprehensive REST API with 107+ endpoints
    covering the complete application lifecycle.

    ## Authentication

    Most endpoints require JWT authentication. There are two authentication modes:

    - **OIDC Mode** (Production): Uses Janua SSO with RS256 JWT validation
    - **Local Mode** (Development): Uses self-signed JWT with password-based auth

    Include the token in the `Authorization` header as `Bearer <token>`.

    ## Rate Limiting

    - Auth endpoints: 10 requests/minute (5 for register/login)
    - Standard endpoints: 100 requests/minute
    - Build endpoints: 20 requests/minute

    ## Environments

    Services can be deployed to multiple environments:
    - `dev` - Development (default)
    - `staging` - Pre-production testing
    - `prod` - Production

    ## WebSocket Endpoints

    Real-time log streaming is available via WebSocket:
    - `ws://api.enclii.dev/v1/services/{id}/logs/stream` - Service logs
    - `ws://api.enclii.dev/v1/deployments/{id}/logs/stream` - Deployment logs
    - `ws://api.enclii.dev/v1/services/{id}/builds/{build_id}/logs/stream` - Build logs

  version: 1.0.0
  contact:
    name: MADFAM Engineering
...
servers:
...
paths:
...
  /v1/projects/{slug}/cron-jobs:
    post:
      summary: Create cron job
      description: Creates a recurring scheduled job for a service
      tags: [timetable]
      operationId: createCronJob
      parameters:
        - $ref: '#/components/parameters/projectSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, name, schedule, command]
              properties:
                service_id:
                  type: string
                  format: uuid
                name:
                  type: string
                schedule:
                  type: string
                  description: "5-field cron expression (e.g., '*/5 * * * *')"
                command:
                  type: string
                image:
                  type: string
                timeout:
                  type: integer
                  default: 300
                retries:
                  type: integer
                  default: 0
                concurrency:
                  type: string
                  enum: [allow, forbid, replace]
                  default: forbid
      responses:
        '201':
          description: Cron job created
        '400':
          description: Invalid request
        '404':
          description: Project not found
        '409':
          description: Cron job with this name already exists
    get:
      summary: List cron jobs
      description: Lists all cron jobs for a project
      tags: [timetable]
      operationId: listCronJobs
      parameters:
        - $ref: '#/components/parameters/projectSlug'
      responses:
        '200':
          description: List of cron jobs
  /v1/cron-jobs/{id}:
    get:
      summary: Get cron job
      tags: [timetable]
      operationId: getCronJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cron job details
        '404':
          description: Cron job not found
    patch:
      summary: Update cron job
      tags: [timetable]
      operationId: updateCronJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                schedule:
                  type: string
                command:
                  type: string
                image:
                  type: string
                timeout:
                  type: integer
                retries:
                  type: integer
                suspended:
                  type: boolean
                concurrency:
                  type: string
                  enum: [allow, forbid, replace]
      responses:
        '200':
          description: Cron job updated
        '404':
          description: Cron job not found
    delete:
      summary: Delete cron job
      tags: [timetable]
      operationId: deleteCronJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cron job deleted
        '404':
          description: Cron job not found
  /v1/cron-jobs/{id}/runs:
    get:
      summary: List cron job runs
      description: Lists execution history for a cron job
      tags: [timetable]
      operationId: listCronJobRuns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of cron job runs
        '404':
          description: Cron job not found
  /v1/projects/{slug}/one-off-jobs:
    post:
      summary: Create one-off job
      description: Creates a single-execution job for a service
      tags: [timetable]
      operationId: createOneOffJob
      parameters:
        - $ref: '#/components/parameters/projectSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, name, command]
              properties:
                service_id:
                  type: string
                  format: uuid
                name:
                  type: string
                command:
                  type: string
                image:
                  type: string
                timeout:
                  type: integer
                  default: 300
                run_at:
                  type: string
                  format: date-time
                  description: "RFC3339 timestamp. Omit to run immediately."
      responses:
        '201':
          description: One-off job created
        '400':
          description: Invalid request
        '404':
          description: Project not found

  # ============================================
  # JUNCTION ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or OIDC callback
  parameters:
...
    projectSlug:
      name: slug
      in: path
      required: true
      description: Project slug
      schema:
        type: string
...
security:
  - bearerAuth: []